+++ src/sc/photogallery/browser/photogallery.py
@@ -25,13 +25,27 @@ class View(grok.View):
     def image(self, obj, scale='large'):
         """Return an image scale if the item has an image field.
 
-        :param item: [required]
-        :type item: content type object
+        :param obj: [required]
+        :type obj: content type object
         :param scale: the scale to be used
         :type scale: string
         """
         scales = obj.restrictedTraverse('@@images')
         return scales.scale('image', scale)
+
+    def description(self, obj):
+        """Return the description itens of the tile.
+
+        :param obj: [required]
+        :type obj: content type object
+        """
+        slide_num = self.results().index(obj) + 1
+        slide_count = len(self.results())
+        return {
+            'slide_num': slide_num,
+            'slide_count': slide_count,
+            'description': obj.Description() or obj.Title()
+        }
 
     def localized_time(self, obj, long_format=False):
         """Return the object time in a user-friendly way.