+++ src/sc/photogallery/tests/test_upgrades.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+from sc.photogallery.testing import INTEGRATION_TESTING
+
+import unittest
+
+
+class BaseUpgradeTestCase(unittest.TestCase):
+
+    layer = INTEGRATION_TESTING
+
+    def setUp(self, from_version, to_version):
+        self.portal = self.layer['portal']
+        self.setup = self.portal['portal_setup']
+        self.profile_id = u'sc.photogallery:default'
+        self.from_version = from_version
+        self.to_version = to_version
+
+    def _upgrades_to_do(self):
+        self.setup.setLastVersionForProfile(self.profile_id, self.from_version)
+        upgrades = self.setup.listUpgrades(self.profile_id)
+        assert len(upgrades) > 0
+        return len(upgrades[0])
+
+
+class To1001TestCase(BaseUpgradeTestCase):
+
+    def setUp(self):
+        BaseUpgradeTestCase.setUp(self, u'1000', u'1001')
+
+    def test_upgrade_to_1001_registrations(self):
+        version = self.setup.getLastVersionForProfile(self.profile_id)[0]
+        self.assertTrue(version >= self.to_version)
+        self.assertEqual(self._upgrades_to_do(), 1)