+++ CHANGES.rst
@@ -3,7 +3,7 @@ Changelog
 
 There's a frood who really knows where his towel is.
 
-1.0a2 (unreleased)
+1.0b1 (2016-03-15)
 ------------------
 
 - Remove hard dependency on plone.app.referenceablebehavior as Archetypes is no longer the default framework in Plone 5.